requirements.txt
asgiref==3.9.2
Django==5.2.6
django-extensions==4.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
pillow==11.3.0
psycopg2-binary==2.9.11
sqlparse==0.5.3
tzdata==2025.2
